#16 : prepare Import; change any "Points" to double * Participant: add AlreadyPromotedA/B if "import says this" * any "Points": change to double * ImportTypeEnum: move to "DanceCompetitionHelper.Database/Enum" * IDanceCompetitionHelper: add "ImportOrUpdateCompetition" * Views/Participant/Index.cshtml: highlight "already promoted" * OetsvCompetitionImporter: read/mind "Org Class" to check if "already promoted"

+++ src/DanceCompetitionHelper.Database/Tables/CompetitionClassHistory.cs
@@ -61,11 +61,11 @@ namespace DanceCompetitionHelper.Database.Tables
 
         [Range(0, int.MaxValue)]
         public int MinStartsForPromotion { get; set; }
-        [Range(0, int.MaxValue)]
-        public int MinPointsForPromotion { get; set; }
+        [Range(0, double.MaxValue)]
+        public double MinPointsForPromotion { get; set; }
 
-        [Range(0, int.MaxValue)]
-        public int PointsForFirst { get; set; }
+        [Range(0, double.MaxValue)]
+        public double PointsForFirst { get; set; }
 
         [Range(0, int.MaxValue)]
         public int ExtraManualStarter { get; set; }
