+++ src/DanceCompetitionHelper.Web/Models/CompetitionClassModels/CompetitionClassViewModel.cs
@@ -64,12 +64,12 @@ namespace DanceCompetitionHelper.Web.Models.CompetitionClassModels
         public int MinStartsForPromotion { get; set; }
 
         [FromForm]
-        [Range(0, int.MaxValue)]
-        public int MinPointsForPromotion { get; set; }
+        [Range(0, double.MaxValue)]
+        public double MinPointsForPromotion { get; set; }
 
         [FromForm]
-        [Range(0, int.MaxValue)]
-        public int PointsForFirst { get; set; }
+        [Range(0, double.MaxValue)]
+        public double PointsForFirst { get; set; }
 
         [FromForm]
         [Range(0, int.MaxValue)]
